Added registration for patients

+++ final-capstone/java/src/main/java/com/techelevator/controller/AuthenticationController.java
@@ -54,7 +54,7 @@ public class AuthenticationController {
 
     @ResponseStatus(HttpStatus.CREATED)
     @RequestMapping(value = "/register", method = RequestMethod.POST)
-    public Long register(@Valid @RequestBody RegisterUserDTO newUser) {
+    public int register(@Valid @RequestBody RegisterUserDTO newUser) {
         try {
             User user = userDAO.findByUsername(newUser.getUsername());
             throw new UserAlreadyExistsException();
